requirements.txt
anyio==4.0.0
asttokens==2.4.0
backcall==0.2.0
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
comm==0.1.4
debugpy==1.8.0
decorator==5.1.1
executing==2.0.0
Flask==3.0.0
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
idna==3.4
ipython==8.16.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
jupyter_client==8.3.1
jupyter_core==5.3.2
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
nest-asyncio==1.5.8
packaging==23.2
parso==0.8.3
pickleshare==0.7.5
platformdirs==3.11.0
prompt-toolkit==3.0.39
psutil==5.9.5
pure-eval==0.2.2
Pygments==2.16.1
python-dateutil==2.8.2
python-telegram-bot==20.6
pywin32==306
pyzmq==25.1.1
requests==2.31.0
six==1.16.0
sniffio==1.3.0
stack-data==0.6.3
tornado==6.3.3
traitlets==5.11.2
urllib3==2.0.6
wcwidth==0.2.8
Werkzeug==3.0.0
